chatbot_backend/src/prompt_template.py
from langchain.prompts import ChatPromptTemplate

### Normal RAG Prompt Template to provide the answer based on the provide context ###
rag_template = """You are an AI assistant that ONLY answers questions based on the provided context.

INSTRUCTIONS:
1. ONLY use information explicitly stated in the source context delimited by triple hashes (###)
2. NEVER use your external knowledge or information outside the provided context
3. If the context doesn't contain information to answer the question, respond with ONLY:
   "I don't have information about this in the provided documents."
4. DO NOT GUESS or MAKE UP information that isn't explicitly in the context
5. Include citation ids in the answer to verify the information found in the context

Source context: 
###{context}###

User question: {question}

"""

def rag_prompt(input: dict) -> str:
    prompt = ChatPromptTemplate.from_template(rag_template)
    rag_prompt = prompt.invoke(input)
    return rag_prompt

### History RAG Prompt Template to provide the answer based on the provide context and history ###
chat_rag_template = """You are an AI assistant with memory of previous conversations that ONLY answers based on provided context.

INSTRUCTIONS:
1. ONLY use information explicitly stated in the context below
2. NEVER use your external knowledge or information outside the provided context
3. If the context doesn't contain information to answer the question, respond with ONLY:
   "I don't have information about this in the provided documents."
4. DO NOT GUESS or MAKE UP information that isn't explicitly mentioned in the context
5. Include citation ids in the answer to verify the information is derived from the given context
6. Answer to user question in full sentences, similar to how a human would respond
7. If user asks something based on previous conversation, summarize the history and use it to answer the question

Previous conversation summary:
{history}

Source context: 
###{context}###

The user asked: {question}
"""

def chat_rag_prompt(input: dict) -> str:
    prompt = ChatPromptTemplate.from_template(chat_rag_template)
    chat_prompt = prompt.invoke(input)
    return chat_prompt

chat_wiki_template = """You are an AI assistant with memory of previous conversations that ONLY answers based on provided context.

INSTRUCTIONS:
1. ONLY use information explicitly stated in the context below
2. NEVER use your external knowledge or information outside the provided context
3. If the context doesn't contain information to answer the question, respond with ONLY:
   "I don't have information to answer the question based on the provided context."
4. DO NOT GUESS or MAKE UP information that isn't explicitly mentioned in the context
5. Include citation ids in output to verify the information is derived from the given context
6. Answer to user question in full sentences, similar to how a human would respond

Previous conversation summary:
{history}

Source context: 
###{context}###

The user asked: {question}

Output format:
- answer: The answer to the user's question based on the provided context.
- citations: A list of citation ids that support the answer, formatted as a list of integers.
"""

def chat_wiki_prompt(input: dict) -> str:
    prompt = ChatPromptTemplate.from_template(chat_wiki_template)
    chat_prompt = prompt.invoke(input)
    return chat_prompt