chatbot_backend/requirement.txt
fastapi
python-dotenv
langchain-groq
langchain-community
pypdf
langchain-huggingface
faiss-cpu
langchain-chroma
passlib
python-jose
pydantic[email]
python-multipart
pandas
sentence-transformers
wikipedia